requirements.txt
argcomplete==1.9.3
aspy.yaml==1.1.1
base58==1.0.0
boto3==1.9.7
botocore==1.12.7
cached-property==1.5.1
certifi==2018.8.24
cfgv==1.1.0
cfn-flip==1.0.3
chardet==3.0.4
click==6.7
coverage==4.5.1
Django==2.1.1
django-cors-headers==2.4.0
djangorestframework==3.8.2
docutils==0.14
durationpy==0.5
factory-boy==2.11.1
Faker==0.9.1
future==0.16.0
hjson==3.0.1
identify==1.1.6
idna==2.7
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
nodeenv==1.3.2
pep8==1.7.1
placebo==0.8.2
pre-commit==1.11.0
psycopg2-binary==2.7.5
pycodestyle==2.4.0
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2018.5
PyYAML==3.12
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
text-unidecode==1.2
toml==0.9.6
tqdm==4.19.1
troposphere==2.3.3
Unidecode==1.0.22
urllib3==1.23
virtualenv==16.0.0
Werkzeug==0.14.1
wsgi-request-logger==0.4.6
zappa==0.46.2
